# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.ErrorResp:
    properties:
      message:
        type: string
    type: object
...
  dto.ApiError:
    properties:
      code:
        type: string
      details:
        type: string
      stack:
        description: Hide in production
    type: object
...
  dto.ApiResponse-array_dto_CartDiscount:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CartDiscount'
        type: array
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.MetaInfo'
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
...
  dto.ApiResponse-dto_CartDetail:
    properties:
      data:
        $ref: '#/definitions/dto.CartDetail'
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.MetaInfo'
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
...
  dto.ApiResponse-dto_PaymentResult:
    properties:
      data:
        $ref: '#/definitions/dto.PaymentResult'
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.MetaInfo'
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
...
  dto.ApiResponse-string:
    properties:
      data:
        type: string
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.MetaInfo'
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  dto.AttributeDetail:
    properties:
      id:
        type: integer
      name:
        type: string
      values:
        items:
          $ref: '#/definitions/dto.AttributeValueDetail'
        type: array
    type: object
  dto.AttributeValueDetail:
    properties:
      id:
        type: integer
      name:
        type: string
      value:
        type: string
    type: object
...
  dto.CartDetail:
    properties:
      cartItems:
        items:
          $ref: '#/definitions/dto.CartItemDetail'
        type: array
      createdAt:
        type: string
      discountAmount:
        type: number
      id:
        type: string
      totalPrice:
        type: number
      updatedAt:
        type: string
    type: object
  dto.CartDiscount:
    properties:
      code:
        type: string
      description:
        type: string
      discountType:
        type: string
      discountValue:
        type: number
      id:
        type: string
      isStackable:
        type: boolean
      maxDiscountAmount:
        type: number
      minOrderValue:
        type: number
      validFrom:
        type: string
      validUntil:
        type: string
    type: object
  dto.CartItemDetail:
    properties:
      attributes:
        items:
          $ref: '#/definitions/dto.AttributeDetail'
        type: array
      discountAmount:
        type: number
      id:
        type: string
      imageUrl:
        type: string
      name:
        type: string
      price:
        type: number
      productId:
        type: string
      quantity:
        type: integer
      sku:
        type: string
      stock:
        type: integer
      variantId:
        type: string
    required:
    - id
    - productId
    - variantId
    type: object
...
  dto.ErrorResp:
    properties:
      error:
        $ref: '#/definitions/dto.ApiError'
    type: object
...
  dto.MetaInfo:
    properties:
      method:
        type: string
      path:
        type: string
      requestId:
        type: string
      timestamp:
        type: string
    type: object
...
  dto.Pagination:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  dto.PaymentResult:
    properties:
      clientSecret:
        type: string
      orderId:
        type: string
      paymentId:
        type: string
      paymentIntentId:
        type: string
      status:
        type: string
      totalPrice:
        type: number
    type: object
...
  models.CheckoutModel:
    properties:
      addressId:
        type: string
      discountCodes:
        items:
          type: string
        type: array
      paymentMethod:
        enum:
        - cod
        - stripe
        type: string
    required:
    - addressId
    - paymentMethod
    type: object
...
  models.UpdateCartItemQtyModel:
    properties:
      quantity:
        type: integer
    required:
    - quantity
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: This is a sample server for a simple e-commerce API.
  title: E-Commerce API
paths:
...
  /carts:
    get:
      consumes:
      - application/json
      description: get cart details by user ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-dto_CartDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResp'
      summary: Get cart details by user ID
      tags:
      - cart
    post:
      consumes:
      - application/json
      description: create a new cart for a user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-dto_CartDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResp'
      summary: Create a new cart
      tags:
      - carts
  /carts/available-discounts:
    get:
      consumes:
      - application/json
      description: get cart discounts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-array_dto_CartDiscount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResp'
      summary: Get cart discounts
      tags:
      - carts
  /carts/checkout:
    post:
      consumes:
      - application/json
      description: update product items in the cart
      parameters:
      - description: Checkout input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CheckoutModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-dto_PaymentResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResp'
      summary: Update product items in the cart
      tags:
      - carts
  /carts/clear:
    put:
      consumes:
      - application/json
      description: clear the cart
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResp'
      summary: Clear the cart
      tags:
      - carts
  /carts/items/{id}:
    delete:
      consumes:
      - application/json
      description: remove a product from the cart
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResp'
      summary: Remove a product from the cart
      tags:
      - carts
  /carts/items/{variant_id}:
    post:
      consumes:
      - application/json
      description: add a product to the cart
      parameters:
      - description: Add product to cart input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCartItemQtyModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResp'
      summary: update product quantity in the cart
      tags:
      - carts
...
swagger: "2.0"